Rearrange shorthand module
